Move file_cache into src/cache/ and update requires
